requirements.txt
Authlib==1.3.2
fer==22.4.0
Flask==3.0.3
numpy==1.25.1
opencv_contrib_python==4.8.1.78
opencv_python==4.5.5.64
python-dotenv==1.0.1
transformers==4.31.0
